add method to import sample case data
