+++ neuralqa/utils/__init__.py
@@ -1,2 +1,3 @@
 from .config_utils import ConfigParser
 from .file_utils import *
+from .data_utils import import_case_data
